third_party/blink/web_tests/flag-specific/disable-layout-ng/virtual/blink-cors/external/wpt/service-workers/service-worker/registration-schedule-job.https-expected.txt

This is a testharness.js-based test.
FAIL different scriptURL and updateViaCache promise_test: Unhandled rejection with value: object "TypeError: Cannot read property 'scriptURL' of null"
FAIL different type assert_equals: expected (string) "classic" but got (undefined) undefined
Harness: the test ran to completion.
